requirements.txt
ultralytics==8.0.126
opencv-python
pillow
numpy
pandas
tqdm
